requirements.txt
sounddevice==0.4.6
webrtcvad==2.0.10
numpy==1.26.4
pyzmq==25.1.2
nats-py==2.7.2
